refactor(web): extract message-list virtualization
Move MessageList (+ private VirtualizedMessageList/VirtualizedMessageRow
+ scroll/measurement helpers + virtualization constants) into
session-detail/message-list.tsx. It imports the already-extracted
MessageItem. SessionDetail drops to 510. Fourth sub-domain of CS-13.
